Remove MPI and some router dependency
Most MPI code has been removed, but the router is still very tightly coupled
with the rest of the code.  Right now, compiled meld program runs without
visible error, but further test need to be conducted and router and remote
can be further decoupled from the rest of the program.